+++ components-ui/app/src/tol-ui/src/general/Tag.tsx
@@ -6,7 +6,7 @@ SPDX-License-Identifier: MIT
 
 import { ReactNode } from "react";
 
-import { LIGHT_COLOURS, TLightColour } from "..";
+import { HoverOverlay, Icon, LIGHT_COLOURS, TLightColour } from "..";
 import type { TColour } from "..";
 
 
@@ -23,22 +23,37 @@ export interface PTag {
    * CSS class name to allow for additional styles
    */
   className?: string;
+  /**
+   * Optional FontAwesome icon name to display before the tag content.
+   */
+  icon?: string;
+  /**
+   * Optional text to display in a tooltip when the tag is hovered.
+   */
+  tooltip?: string;
 }
 
 /**
  * A simple tag component, can be overidden with a different colour type
  */
 export function Tag(props: PTag) {
-  const { children, type, className } = props;
+  const { children, type, className, icon, tooltip } = props;
 
   const style = type ? ({
     backgroundColor: `var(--tol-${type})`,
     color: `var(--tol-${Object.values(LIGHT_COLOURS).includes(type as TLightColour) ? "text" : "light"})`
   }) : undefined;
 
-  return (
-    <div className={`tol-tag ${className}`} style={style}>
+  const tag = (
+    <div className={`tol-tag${className ? ` ${className}` : ""}`} style={style}>
+      {icon && <Icon icon={icon} />}
       {children}
     </div>
   );
+
+  return tooltip ? (
+    <HoverOverlay contents={tooltip}>
+      {tag}
+    </HoverOverlay>
+  ) : tag;
 }